Use Node 20 for development (#12)
We are using the Octokit SDK to interface with the REST and GraphQL
GitHub APIs. As this SDK makes HTTP requests, if it cannot find a
`fetch` function, it will ask you to provide one. Because we're using a
version of Node that does not ship with the Fetch API, we have to get
`fetch` from `node-fetch` and also bring in `@types/node-fetch`.

This is a poor solution. Not only does importing `fetch` violate the
lint rule `@typescript-eslint/no-shadow` (because ESLint thinks `fetch`
is already a global), it also doesn't work, as one of the libraries in
the Octokit SDK assumes that the types `Request` and `Response` are
available globally in TypeScript-land (when they are not).

There is no reason to use an older version of Node for this repo at the
moment, as it's only designed to be run locally. We also don't need to
test against other Node versions either. So this commit upgrades Node to
20 and adjusts the GitHub workflow to only test on whatever version is
set in `.nvmrc`.