Improve contributor docs refs count script (#17)
The script to count the number of references to the contributor-docs
throughout comments in MetaMask repositories has never provided accurate
counts. Upon reviewing this script, I noticed that while comments posted
directly to pull requests were included in the count, comments posted as
a part of reviews, which are separate types in the GitHub GraphQL API,
were excluded. This resulted in a smaller count than expected.

In addition, the cutoff for data that the script used — 5000 maximum
comments per repo — hasn't been particularly useful. We need to define
_some_ kind of cutoff because it would be impractical to pull all
comments from all time, but since we use this script to gauge progress
toward OKRs, it helps us better to define a fixed target, in this case
4/1/2023, the start of the first quarter where the Shared Libraries team
began working on the contributor docs. In addition, the output of this
script now breaks down the counts per quarter instead of giving a lump
sum. Here is an example:

```
About to retrieve data. Please be patient, this could take a while:

✔ Retrieving data for core since 2023-04-01T00:00:00.000Z
✔ Retrieving data for design-tokens since 2023-04-01T00:00:00.000Z
✔ Retrieving data for metamask-extension since 2023-04-01T00:00:00.000Z
✔ Retrieving data for metamask-mobile since 2023-04-01T00:00:00.000Z
✔ Retrieving data for snaps since 2023-04-01T00:00:00.000Z

----------------------

Number of references to contributor-docs by repository:

- core
  - Q2-2023: 0
  - Q3-2023: 0
  - Q4-2023: 2
  - Q1-2024: 1
- design-tokens
  - Q2-2023: 0
  - Q3-2023: 0
  - Q4-2023: 0
  - Q1-2024: 0
- metamask-extension
  - Q2-2023: 0
  - Q3-2023: 0
  - Q4-2023: 0
  - Q1-2024: 2
- metamask-mobile
  - Q2-2023: 0
  - Q3-2023: 0
  - Q4-2023: 0
  - Q1-2024: 4
- snaps
  - Q2-2023: 0
  - Q3-2023: 0
  - Q4-2023: 0
  - Q1-2024: 0

Total number of references: 9
```